Start building parallel installable 2.0 api only
Since we are getting towards a complete Cogl 2.0 api this breaks 1.x api
compatibility by no longer building a libcogl.so and instead building a
libcogl2.so library.

The headers are now installed into an include/cogl2 sub-directory that
wont conflict with an installed cogl 1.x package.

The cogl-2.0-experimental.pc pkg-config file is now named
cogl2.pc and similarly the reference manual has been renamed from
cogl-2.0-experimental to cogl2.

The examples have not been namespaced and so they are not parallel
installable.

Since we haven't yet reached the 2.0 api yet the version has been bumped
to 1.99.1

All of the conformance tests have been updated so they no longer test
the 1.x apis, and since I was going through all the tests I took the
opportunity to rename the global ctx and fb variables to test_ctx and
test_fb.

Reviewed-by: Neil Roberts <neil@linux.intel.com>