Revert "Add compatibility API for functions removed after 1.9.8"
This reverts commit 361bd516f3d678d6d8ec08c1f01cf872e10a4f43.

This patch was only added to provide compatibility with the 1.9.8
release after an API cleanup removed some redundant experimental
symbols. This is no longer needed because the master branch is now
able to break API again.

I've also removed some later changes that were added to reflect the
patch in the MSVC project files.

Conflicts:

	cogl/Makefile.am

Reviewed-by: Robert Bragg <robert@linux.intel.com>