mesh-renderer: Use a custom uniform for the normal matrix
Previously the lighting for the mesh renderer was using
gl_NormalMatrix to transform the normals. This builtin is not
available on GLES2 so instead it now creates its own uniform which
gets updated every time the mesh is rendered. The matrix is calculated
by grabbing the current modelview matrix and calculating its inverse
transpose.