Introduces ray picking
One of the methods to pick 3d objects is to intersect the picking ray
introduced in an earlier commit with the geometry of the scene.

To do so, we cycle through the pickable entities, grab their vertex data
and intersect the ray with that vertex data. The closer object we
intersected is our picked object.