Fix the forwarding of mouse wheel events
Mousewheel events are a bit special in SDL and don't leave state, so the
return of GetMouseState() is 0 for them. Instead SDL generates a DOWN
and UP events. We act on those to set the state.

Take the opportunity to also give better names than 4 and 5 for the
mouse wheel events.