Extract the include path for android_native_app_glue from the module
This adds an m4 macro to check for an Android module by calling
ndk-build with a dummy module that includes it. If it is found then it
will set a CFLAGS variable to the include path for the module. The
configure script now uses this to get the right include path for
android_native_app_glue.