Call the shell init and fini cb around each main loop implementation
On Android, the context needs to be created after the init function
has been called because Cogl needs to know the native window first.
The init callback is called in response to the APP_CMD_INIT_WINDOW
command so it works to try to create the context in the init function
in that case. However when building for SDL the init callback is
called when the shell is initialised and this only happens once the
context is created so it previously wasn't possible to create the
context in response to the init event. This patch changes it so the
init and fini callbacks are called before and after starting the main
loop in SDL and glib so that the application can work the same in all
three cases.

The editor code has been updated to create the context in the init
callback.