Replace _FLAG macros with bitfields
I think it's probably a personal taste thing but I've always found
XYZ_{HAS,SET,CLEAR}_FLAG macros confusing. They are only barely shorter
to type than using the bitwise operators but imho they are less readable
for several reasons, including not being able to use ctags to jump to
the FLAGS enum and leading me to doubt (and then have to check) what
names to use for them - should it be _GET_FLAG() or _HAS_FLAG() and
should it be _CLEAR_FLAG() or _UNSET_FLAG()?

This patch removes all the boilerplate _FLAG macros and flag enums and
instead adds corresponding bitfields which I think is the best of both
worlds since the code is readable and it's even less typing.