Remove columnised formatting of prototypes
This is a personal taste thing but I tend to prefer the prototype style
be consistent with the C code, for it not to force most lines over 80
chars and to not have to keep maintaining the indentation across a whole
file just to add a single new prototype which then makes patches that
others have to review contain superfluous noise. Another good reason to
avoid columnising code imho is to make automatic code refactoring much
easier. For example in Cogl I've been thankful several times that we
don't columnise code for aesthetics, since that means it's easier to run
a sed scripts to rename types for example without having to then
manually audit all of the changes to make indentation fixes.