Reorganize headers to help avoid circular dependencies
Since we want to make entities and components into RigObjects we need to
re-jig things a bit to avoid a circular dependency in the header.

This splits out all the api definitions from rig.h into a separate
rig-context.h so rig.h is now just an aggregation header that pulls
together all the apis that can be consumed by an editor built on top of
rig. rig.h should no longer need to be used internally (Except for
in rig-shell.{c,h} as a special case until we add a config.h as rig.h
defines what platform support is enabled).

Part of the problem before was that there were various non-opaque types
defined in rig.h after all the header includes and so if any of those
included headers needed access to one of the non-opaque types they would
be out of luck because the _RIG_CONTEXT_H_ guard would already be
defined so a further include of rig.h wouldn't work.