editor: remove pivot, fix transformed camera arcball
This removes the pivot entity which was used for transforming the world
around for navigation and instead we can now move the camera entity
around. This patch also fixes how the arcball rotations modify the
entities to take into account the current viewing rotation so now the
arcball should behave well from all angles.