property: Add a convenience function to link two properties
This adds rig_property_set_copy_binding to set the binding on a given
property to be automatically updated based on the value of the source
property. This will probably be quite a common operation (for example,
linking the value of a slider to a property) so it makes sense to have
a convenience function.

The binding uses another convenience function which has been made
public called rig_property_copy_value which copies the value of one
property to another.