number-slider: Add a text control when the slider is clicked
When the middle of the slider is clicked without moving the mouse it
now changes to show a text control to directly edit the number.
Activating the text control or clicking anywhere outside of the slider
causes the text to be committed. Pressing escape causes the text to be
cancelled.