Add the start of a simple number slider widget
This adds a graphable/paintable object to represent a control for
editing a widget. The control is inspired by the ones from Blender.
The control displays a label along with the slider value directly in a
button. On either side of the button is an arrow to increment or
decrement the value by one step. The size of the step can be
configured. The value can also be adjusted by clicking on the button
and dragging the mouse left or right to get a live view of how the
change affects the scene.

The plan is to also add the ability to edit the value directly with a
text control if is clicked on without dragging and also to make moving
the cursor up or down while dragging increase the scale to change the
value faster.