Allow multiple input grabs and require a camera
There can now be multiple input grabs added with rig_shell_grab_input.
The events are given to all grabs in reverse order until one of the
callbacks returns RIG_INPUT_EVENT_STATUS_HANDLED.

The grab can also optionally be given a camera. When the event is
passed to the grab this camera will temporarily be set on the event.
This is useful because usually a grab is set up in response to an
input event from an input region and in that case there will be a
particular camera that the event is relative to. Usually the grab
handler will want to interpret the event relative to this camera too.
This also fixes a crash in RigText where the grab for dragging the
selection expects a camera to be available but it wasn't.