text: Ungrab the key focus on destruction
Previously if the text control was destroyed while it had focus then
when a new widget tries to grab focus the shell will call the focus
lost callback with an invalid pointer it would segfault.

The unfocusing is acheived with a new shell function called
rig_shell_ungrab_key_focus.