text: Add a public function to grab the focus
This adds a rig_text_grab_focus function to force the text control to
grab the focus. This is useful for example when first displaying a set
of text controls to determine which should have the focus.

In the long run it might be nicer to have an interface for focusable
objects so that the text control could react to gaining the focus
instead of having to actively grab the focus itself. That would help
to implement a top level focus manager.