Move unproject event utility function from RigNumberSlider to RigShell
RigNumberSlider had a little utility function to unproject the
coordinates of a motion event into the coordinate space of the given
graphable object. I think this could be quite a useful general purpose
utility so this patch moves it to rig-shell and changes the name to
rig_motion_event_unproject().