Add a drop down selector widget
This adds a widget to select from a list of values. Each value is a
pair containing the name of the value and an integer constant. The
value property of the control is the integer constant of the currently
selected item. The value pairs can be set with a function taking an
array or varargs.